refactor(app): split app.rs
